Revert "解决开启事务失败时, 引用没有回收的问题。"
This reverts commit a4d11097d024b8d43f2efe459e21671f2caddde2.
